requirements-render.txt
# Render-specific requirements - full set for complete functionality
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.1
sqlalchemy==2.0.23
requests==2.31.0
openai==1.3.5
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.5
langchain-community==0.0.5
sentence-transformers==2.2.2
stable-baselines3==2.1.0
gymnasium==0.29.1
torch==2.1.2
pandas==2.0.3
scikit-learn==1.3.0
streamlit==1.28.0
plotly==5.18.0
pillow==10.1.0
pytesseract==0.3.10
pymupdf==1.23.21
numpy-stl==3.1.0
pytest==7.4.3
